Fixing some pointer to integer casts by using uintptr_t.

+++ src/nvme_cmds.c
@@ -216,7 +216,7 @@ int nvme_Security_Send(tDevice *device, uint8_t securityProtocol, uint16_t secur
     adminCommand.commandType = NVM_ADMIN_CMD;
     adminCommand.cmd.adminCmd.opcode = NVME_ADMIN_CMD_SECURITY_SEND;
     adminCommand.commandDirection = XFER_DATA_OUT;
-    adminCommand.cmd.adminCmd.addr = (uint64_t)ptrData;//TODO: does this need a cast?
+    adminCommand.cmd.adminCmd.addr = (uintptr_t)ptrData;//TODO: does this need a cast?
     adminCommand.ptrData = ptrData;
     adminCommand.dataSize = dataLength;
     adminCommand.cmd.adminCmd.cdw10 = M_BytesTo4ByteValue(securityProtocol, M_Word1(securityProtocolSpecific), M_Word0(securityProtocolSpecific), nvmeSecuritySpecificField);
@@ -243,7 +243,7 @@ int nvme_Security_Receive(tDevice *device, uint8_t securityProtocol, uint16_t se
     adminCommand.commandType = NVM_ADMIN_CMD;
     adminCommand.cmd.adminCmd.opcode = NVME_ADMIN_CMD_SECURITY_RECV;
     adminCommand.commandDirection = XFER_DATA_IN;
-    adminCommand.cmd.adminCmd.addr = (uint64_t)ptrData;//TODO: does this need a cast?
+    adminCommand.cmd.adminCmd.addr = (uintptr_t)ptrData;//TODO: does this need a cast?
     adminCommand.ptrData = ptrData;
     adminCommand.dataSize = dataLength;
     adminCommand.cmd.adminCmd.cdw10 = M_BytesTo4ByteValue(securityProtocol, M_Word1(securityProtocolSpecific), M_Word0(securityProtocolSpecific), nvmeSecuritySpecificField);
@@ -483,7 +483,7 @@ int nvme_Firmware_Image_Dl(tDevice *device,\
     ImageDl.cmd.adminCmd.opcode = NVME_ADMIN_CMD_DOWNLOAD_FW;
     ImageDl.commandType = NVM_ADMIN_CMD;
     ImageDl.commandDirection = XFER_DATA_OUT;
-    ImageDl.cmd.adminCmd.addr = (uint64_t)ptrData;
+    ImageDl.cmd.adminCmd.addr = (uintptr_t)ptrData;
     ImageDl.ptrData = ptrData;
     ImageDl.dataSize = numberOfBytes;
     ImageDl.cmd.adminCmd.cdw10 = (numberOfBytes >> 2) - 1; //Since this is, 0 based, number of DWords not Bytes. 
@@ -542,7 +542,7 @@ int nvme_Identify(tDevice *device, uint8_t *ptrData, uint32_t nvmeNamespace, uin
     identify.commandType = NVM_ADMIN_CMD;
     identify.commandDirection = XFER_DATA_IN;
     identify.cmd.adminCmd.nsid = nvmeNamespace;
-    identify.cmd.adminCmd.addr = (uint64_t)ptrData;
+    identify.cmd.adminCmd.addr = (uintptr_t)ptrData;
     identify.cmd.adminCmd.cdw10 = cns;
     identify.timeout = 15;
     identify.ptrData = ptrData;
@@ -694,7 +694,7 @@ int nvme_Get_Log_Page(tDevice *device, nvmeGetLogPageCmdOpts * getLogPageCmdOpts
     #if defined(VMK_CROSS_COMP)
     getLogPage.cmd.adminCmd.addr = (uint32_t)getLogPageCmdOpts->addr;
     #else
-    getLogPage.cmd.adminCmd.addr = (uint64_t)getLogPageCmdOpts->addr;
+    getLogPage.cmd.adminCmd.addr = (uintptr_t)getLogPageCmdOpts->addr;
     #endif
     getLogPage.dataSize = getLogPageCmdOpts->dataLen;
     getLogPage.ptrData = getLogPageCmdOpts->addr;
@@ -793,7 +793,7 @@ int nvme_Reservation_Report(tDevice *device, bool extendedDataStructure, uint8_t
     memset(&nvmCmd, 0, sizeof(nvmeCmdCtx));
     nvmCmd.cmd.nvmCmd.opcode = NVME_CMD_RESERVATION_REPORT;
     nvmCmd.cmd.nvmCmd.nsid = device->drive_info.namespaceID;
-    nvmCmd.cmd.nvmCmd.prp1 = (uint64_t)ptrData;
+    nvmCmd.cmd.nvmCmd.prp1 = (uintptr_t)ptrData;
     nvmCmd.commandDirection = XFER_DATA_IN;
     nvmCmd.commandType = NVM_CMD;
     nvmCmd.dataSize = dataSize;
@@ -830,7 +830,7 @@ int nvme_Reservation_Register(tDevice *device, uint8_t changePersistThroughPower
     memset(&nvmCmd, 0, sizeof(nvmeCmdCtx));
     nvmCmd.cmd.nvmCmd.opcode = NVME_CMD_RESERVATION_REGISTER;
     nvmCmd.cmd.nvmCmd.nsid = device->drive_info.namespaceID;
-    nvmCmd.cmd.nvmCmd.prp1 = (uint64_t)ptrData;
+    nvmCmd.cmd.nvmCmd.prp1 = (uintptr_t)ptrData;
     nvmCmd.commandDirection = XFER_DATA_OUT;
     nvmCmd.commandType = NVM_CMD;
     nvmCmd.dataSize = dataSize;
@@ -868,7 +868,7 @@ int nvme_Reservation_Acquire(tDevice *device, uint8_t reservationType, bool igno
     memset(&nvmCmd, 0, sizeof(nvmeCmdCtx));
     nvmCmd.cmd.nvmCmd.opcode = NVME_CMD_RESERVATION_ACQUIRE;
     nvmCmd.cmd.nvmCmd.nsid = device->drive_info.namespaceID;
-    nvmCmd.cmd.nvmCmd.prp1 = (uint64_t)ptrData;
+    nvmCmd.cmd.nvmCmd.prp1 = (uintptr_t)ptrData;
     nvmCmd.commandDirection = XFER_DATA_OUT;
     nvmCmd.commandType = NVM_CMD;
     nvmCmd.dataSize = dataSize;
@@ -906,7 +906,7 @@ int nvme_Reservation_Release(tDevice *device, uint8_t reservationType, bool igno
     memset(&nvmCmd, 0, sizeof(nvmeCmdCtx));
     nvmCmd.cmd.nvmCmd.opcode = NVME_CMD_RESERVATION_RELEASE;
     nvmCmd.cmd.nvmCmd.nsid = device->drive_info.namespaceID;
-    nvmCmd.cmd.nvmCmd.prp1 = (uint64_t)ptrData;
+    nvmCmd.cmd.nvmCmd.prp1 = (uintptr_t)ptrData;
     nvmCmd.commandDirection = XFER_DATA_OUT;
     nvmCmd.commandType = NVM_CMD;
     nvmCmd.dataSize = dataSize;
